allocater/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
confluent-kafka==2.3.0
cryptography==41.0.7
Django==4.2.8
djangorestframework==3.14.0
dnspython==2.4.2
docker==7.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
mysql-connector-python==8.2.0
mysqlclient==2.2.1
packaging==23.2
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
